+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.12
+* `first_value()` and `last_value()` functions are now deprecated and will be removed
+    in a later version. Use instead `array_value_first()` and `array_value_last()`;
 * `first_key()` and `last_key()` functions are now deprecated and will be removed
     in a later version. Use instead `array_key_first()` and `array_key_last()`;
 * `ends_with()` and `starts_with()` functions are now deprecated and will be removed